chore(python, java): Extract common buildpack code
