+++ buildpacks/common/config/func_yaml.go
@@ -1,7 +1,7 @@
 // Copyright 2021-2022 VMware, Inc.
 // SPDX-License-Identifier: BSD-2-Clause
 
-package python
+package config
 
 import (
 	"os"
