+++ buildpacks/java/java/detect.go
@@ -9,6 +9,8 @@ import (
 	"github.com/buildpacks/libcnb"
 	"github.com/paketo-buildpacks/libpak"
 	"github.com/paketo-buildpacks/libpak/bard"
+
+	"kn-fn/buildpacks/config"
 )
 
 type Detect struct {
@@ -26,7 +28,7 @@ func (d Detect) checkConfigs(cr libpak.ConfigurationResolver) bool {
 func (d Detect) Detect(context libcnb.DetectContext) (libcnb.DetectResult, error) {
 	result := libcnb.DetectResult{}
 
-	funcYaml := ParseFuncYaml(context.Application.Path, d.Logger)
+	funcYaml := config.ParseFuncYaml(context.Application.Path, d.Logger)
 
 	cr, err := libpak.NewConfigurationResolver(context.Buildpack, &d.Logger)
 	if err != nil {
