+++ buildpacks/python/cmd/main/main.go
@@ -9,6 +9,7 @@ import (
 	"github.com/paketo-buildpacks/libpak"
 	"github.com/paketo-buildpacks/libpak/bard"
 
+	"kn-fn/buildpacks/command"
 	"kn-fn/python-function-buildpack/python"
 )
 
@@ -16,6 +17,6 @@ func main() {
 	logger := bard.NewLogger(os.Stdout)
 	libpak.Main(
 		python.Detect{Logger: logger},
-		python.Build{Logger: logger, CommandRunner: python.NewDefaultCommandRunner()},
+		python.Build{Logger: logger, CommandRunner: command.NewDefaultRunner()},
 	)
 }
